refactor[#13] : widgets/codeSnippet 생성
- 기존 CodeSnippet 컴포넌트명 CodeBlock 으로 변경
