+++ app/src/widgets/codeSnippet/ui/CodeSnippet.tsx
@@ -1,0 +1,5 @@
+import { CodeBlock } from "@/features/snippet/ui";
+
+export const CodeSnippet = () => {
+  return <CodeBlock />;
+};
